PyTest DQ Framework/requirements.txt
psycopg2-binary~=2.9.10
pandas~=2.2.3
pytest~=8.4.0
pytest-html~=4.1.1
fastparquet



